+++ sdk/clients/protobuf-client/README.md
@@ -1,0 +1,12 @@
+# Azure Web PubSub Protobuf Client
+
+## Overview
+
+The Azure Web PubSub Protobuf Client allows you to communicate with Azure Web PubSub service using protocol buffers for efficient serialization and deserialization. Protobuf offers a compact binary message format that is faster and more efficient than JSON for many use cases.
+
+## Available Implementations
+
+The Azure Web PubSub Protobuf Client is available in multiple languages:
+
+- **C# (.NET)**: This repository contains the [C# implementation](./csharp)
+- **JavaScript**: The JavaScript implementation is available at [azure-sdk-for-js/sdk/web-pubsub/web-pubsub-client-protobuf](https://github.com/Azure/azure-sdk-for-js/tree/main/sdk/web-pubsub/web-pubsub-client-protobuf)
