+++ sdk/clients/protobuf-client/csharp/Azure.Messaging.WebPubSub.Client.Protobuf.sln
