+++ sdk/clients/protobuf-client/csharp/samples/WebPubSubProtobufSample/WebPubSubProtobufSample.csproj
