+++ sdk/clients/protobuf-client/csharp/tests/Azure.Messaging.WebPubSub.Client.Protobuf.Tests.csproj
